WIP: Few updates for xchain-terra
