+++ packages/xchain-terra/CHANGELOG.md
@@ -1,3 +1,15 @@
+# v.0.1.0-alpha.4 (2022-xx-xx)
+
+## Add
+
+- Helper `getDefaultClientConfig`
+- Helper `getTerraChains`
+- Helper `getDefaultRootDerivationPaths`
+
+## Breaking change
+
+- Extract client related types from `client` to `types/client` (incl. some renaming)
+
 # v.0.1.0-alpha.3 (2022-03-28)
 
 ## FIX
