+++ packages/xchain-terra/src/index.ts
@@ -1,3 +1,4 @@
 export * from './client'
 export * from './const'
 export * from './util'
+export * from './types'
