+++ packages/xchain-terra/src/types/index.ts
@@ -1,0 +1,2 @@
+export * from './client'
+export * from './terra'
